refactor: properties를 config 폴더 안에 정리
