+++ backend/api/src/main/java/com/yat2/episode/auth/oauth/KakaoOAuthClient.java
@@ -1,5 +1,6 @@
 package com.yat2.episode.auth.oauth;
 
+import com.yat2.episode.auth.config.KakaoProperties;
 import com.yat2.episode.auth.dto.KakaoTokenResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.MediaType;
